+++ bin/format_coverage.dart
@@ -19,6 +19,7 @@ class Environment {
     required this.lcov,
     required this.output,
     required this.packagesPath,
+    required this.packagePath,
     required this.prettyPrint,
     required this.prettyPrintFunc,
     required this.prettyPrintBranch,
@@ -36,6 +37,7 @@ class Environment {
   bool lcov;
   IOSink output;
   String? packagesPath;
+  String packagePath;
   bool prettyPrint;
   bool prettyPrintFunc;
   bool prettyPrintBranch;
@@ -54,7 +56,8 @@ Future<void> main(List<String> arguments) async {
     print('  # files: ${files.length}');
     print('  # workers: ${env.workers}');
     print('  sdk-root: ${env.sdkRoot}');
-    print('  package-spec: ${env.packagesPath}');
+    print('  package-path: ${env.packagePath}');
+    print('  packages-path: ${env.packagesPath}');
     print('  report-on: ${env.reportOn}');
     print('  check-ignore: ${env.checkIgnore}');
   }
@@ -63,7 +66,9 @@ Future<void> main(List<String> arguments) async {
   final hitmap = await HitMap.parseFiles(
     files,
     checkIgnoredLines: env.checkIgnore,
+    // ignore: deprecated_member_use_from_same_package
     packagesPath: env.packagesPath,
+    packagePath: env.packagePath,
   );
 
   // All workers are done. Process the data.
@@ -74,7 +79,11 @@ Future<void> main(List<String> arguments) async {
   String output;
   final resolver = env.bazel
       ? BazelResolver(workspacePath: env.bazelWorkspace)
-      : Resolver(packagesPath: env.packagesPath, sdkRoot: env.sdkRoot);
+      : await Resolver.create(
+          packagesPath: env.packagesPath,
+          packagePath: env.packagePath,
+          sdkRoot: env.sdkRoot,
+        );
   final loader = Loader();
   if (env.prettyPrint) {
     output = await hitmap.prettyPrint(resolver, loader,
@@ -116,7 +125,10 @@ Environment parseArgs(List<String> arguments) {
   final parser = ArgParser();
 
   parser.addOption('sdk-root', abbr: 's', help: 'path to the SDK root');
-  parser.addOption('packages', help: 'path to the package spec file');
+  parser.addOption('packages',
+      help: '[DEPRECATED] path to the package spec file');
+  parser.addOption('package',
+      help: 'root directory of the package', defaultsTo: '.');
   parser.addOption('in', abbr: 'i', help: 'input(s): may be file or directory');
   parser.addOption('out',
       abbr: 'o', defaultsTo: 'stdout', help: 'output: may be file or stdout');
@@ -191,6 +203,11 @@ Environment parseArgs(List<String> arguments) {
     }
   }
 
+  final packagePath = args['package'] as String;
+  if (!FileSystemEntity.isDirectorySync(packagePath)) {
+    fail('Package spec "${args["package"]}" not found, or not a directory.');
+  }
+
   if (args['in'] == null) fail('No input files given.');
   final input = p.absolute(p.normalize(args['in'] as String));
   if (!FileSystemEntity.isDirectorySync(input) &&
@@ -254,6 +271,7 @@ Environment parseArgs(List<String> arguments) {
       lcov: lcov,
       output: output,
       packagesPath: packagesPath,
+      packagePath: packagePath,
       prettyPrint: prettyPrint,
       prettyPrintFunc: prettyPrintFunc,
       prettyPrintBranch: prettyPrintBranch,