+++ tool/test_and_collect.sh
@@ -30,7 +30,6 @@ if [[ $(dart --version 2>&1 ) =~ '(dev)' ]]; then
     --lcov \
     --in=var/coverage.json \
     --out=var/lcov.info \
-    --packages=.dart_tool/package_config.json \
     --report-on=lib \
     --check-ignore
 fi